backend/pyproject.toml
[project]
name = "talk-video-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.40.64",
    "celery>=5.5.3",
    "coverage>=7.11.3",
    "cryptography>=46.0.3",
    "dj-database-url>=3.0.1",
    "django-anymail>=13.1",
    "django-cors-headers>=4.9.0",
    "django-storages>=1.14.6",
    "django>=5.2.7",
    "djangorestframework-simplejwt>=5.5.1",
    "djangorestframework>=3.16.1",
    "drf-spectacular>=0.29.0",
    "gunicorn>=23.0.0",
    "langchain-openai>=1.0.1",
    "langchain-postgres>=0.0.16",
    "langchain>=1.0.2",
    "numpy>=2.0.0",
    "openai>=2.6.1",
    "pgvector>=0.3.0",
    "psycopg2-binary>=2.9.11",
    "redis>=7.0.0",
    "scikit-learn>=1.7.2",
    "uvicorn-worker>=0.4.0",
    "uvicorn>=0.38.0"
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "isort>=7.0.0",
]
